files/dfld_box/requirements.txt
influxdb==5.3.2
pyserial==3.5
numpy==2.2.3
smbus3==0.5.5
RPi.bme280==0.2.4
